+++ extensions/com_pungamail/administrator/components/com_pungamail/tmpl/subscribers/default.php
@@ -19,7 +19,8 @@ $statusLabels = [
 	SubscriberRepository::STATUS_UNSUBSCRIBED => Text::_('COM_PUNGAMAIL_SUBSCRIBER_UNSUBSCRIBED'),
 ];
 ?>
-<form action="<?php echo Route::_('index.php?option=com_pungamail&view=subscribers'); ?>" method="post" name="adminForm" id="adminForm">
+<?php echo \Joomla\CMS\Layout\LayoutHelper::render('pungamail.section_navigation', ['section' => 'audience', 'active' => 'subscribers'], JPATH_ADMINISTRATOR . '/components/com_pungamail/layouts'); ?>
+<form action="<?php echo Route::_(AdministratorRoute::subscribers()); ?>" method="post" name="adminForm" id="adminForm">
 	<?php echo LayoutHelper::render('joomla.searchtools.default', ['view' => $this]); ?>
 	<p class="small text-muted mb-3"><?php echo Text::_('COM_PUNGAMAIL_SUPPRESSION_HELP'); ?></p>